Move JS files to src/
